refactor(frontend): move CustomIcon to components
