+++ frontend/.eslintrc.json
@@ -7,6 +7,7 @@
   ],
   "rules": {
     "jsx-a11y/anchor-is-valid": "off",
+    "react-hooks/exhaustive-deps": "warn",
     "no-loop-func": "off",
     "eqeqeq": "off",
     "default-case": "off",
