+++ frontend/src/components/custom-icon.js
