+++ frontend/src/components/search/wiki2-search-result.js
@@ -1,9 +1,9 @@
 import React, { useCallback } from 'react';
 import classNames from 'classnames';
 import PropTypes from 'prop-types';
-import CustomIcon from '../../pages/wiki2/custom-icon';
 import { gettext } from '../../utils/constants';
 import { Utils } from '../../utils/utils';
+import CustomIcon from '../custom-icon';
 import Icon from '../icon';
 
 import './wiki2-search-result.css';
