+++ frontend/src/pages/wiki2/wiki-nav/pages/page-drag-layer.js
@@ -2,9 +2,9 @@ import React from 'react';
 import { useDragLayer } from 'react-dnd';
 import classNames from 'classnames';
 import PropTypes from 'prop-types';
+import CustomIcon from '../../../../components/custom-icon';
 import Icon from '../../../../components/icon';
 import NavItemIcon from '../../common/nav-item-icon';
-import CustomIcon from '../../custom-icon';
 
 const PageDragLayer = ({ pages, getFoldState }) => {
   const { isDragging, item, currentOffset } = useDragLayer((monitor) => ({
