+++ frontend/src/pages/wiki2/wiki-nav/pages/page-item.js
@@ -3,6 +3,7 @@ import { useDrag, useDrop } from 'react-dnd';
 import { getEmptyImage } from 'react-dnd-html5-backend';
 import classnames from 'classnames';
 import PropTypes from 'prop-types';
+import CustomIcon from '../../../../components/custom-icon';
 import { eventBus } from '../../../../components/event-bus';
 import Icon from '../../../../components/icon';
 import OpIcon from '../../../../components/op-icon';
@@ -13,7 +14,6 @@ import wikiAPI from '../../../../utils/wiki-api';
 import NameEditPopover from '../../common/name-edit-popover';
 import NavItemIcon from '../../common/nav-item-icon';
 import { DEFAULT_PAGE_NAME } from '../../constant';
-import CustomIcon from '../../custom-icon';
 import AddNewPageDialog from '../add-new-page-dialog';
 import { INSERT_POSITION } from '../constants';
 import PageDropdownMenu from './page-dropdownmenu';
